pri/requirements.txt
asgiref==3.9.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses==0.6
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djongo==1.2.31
dnspython==2.8.0
idna==3.10
mongoengine==0.29.1
PyJWT==2.10.1
pymongo==4.15.0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
